[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ C++17/백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -1,14 +1,14 @@
 #include <iostream>
 using namespace std;
 
-int atSign(int A, int B)
+long long atSign(long long A, long long B)
 {
     return (A+B)*(A-B);
 }
 
 int main()
 {
-    int A, B;
+    long long A, B;
     cin >> A >> B;
     cout << atSign(A, B);
 }
